Refactor imports to use updated nextui/client-components
