+++ frontend/app/organizations/[org_slug]/tournaments/[tournament_id]/(tournament_page)/@matches/page.tsx
@@ -1,4 +1,4 @@
-import { Card, CardBody, CardHeader } from "@/components/nextui-use-client";
+import { Card, CardBody, CardHeader } from "@/components/nextui/client-components";
 
 interface MatchPageProps {
   params: Promise<{
