+++ frontend/components/cookies.tsx
@@ -3,7 +3,7 @@
 import React from "react";
 import Cookies from "js-cookie";
 
-import { Button, Link } from "@/components/nextui-use-client";
+import { Button, Link } from "@/components/nextui/client-components";
 
 const cookieAttributes = (attrs: Partial<Cookies.CookieAttributes>): Cookies.CookieAttributes => ({
   expires: 7,
