+++ frontend/components/miscellaneous/default-circle.tsx
