+++ frontend/components/miscellaneous/delete.tsx
