+++ frontend/components/navbar/breadcrumbs.tsx
@@ -3,7 +3,7 @@
 import React from "react";
 import { usePathname } from "next/navigation";
 import Link from "next/link";
-import { Breadcrumbs, BreadcrumbItem, NavbarItem } from "@/components/nextui-use-client";
+import { Breadcrumbs, BreadcrumbItem, NavbarItem } from "@/components/nextui/client-components";
 
 export default function BreadcrumbsComponent() {
   const pathname = usePathname();
