+++ frontend/components/navbar/notifications.tsx
@@ -1,4 +1,4 @@
-import { NavbarItem, Popover, PopoverTrigger, Button, Badge, PopoverContent } from "@/components/nextui-use-client";
+import { NavbarItem, Popover, PopoverTrigger, Button, Badge, PopoverContent } from "@/components/nextui/client-components";
 import { auth } from "@clerk/nextjs/server";
 
 import { cn } from "@/lib";
