+++ frontend/components/navbar/settings.tsx
@@ -1,5 +1,5 @@
 import { Icon } from "@iconify/react/dist/iconify.js";
-import { NavbarItem, Button } from "@/components/nextui-use-client";
+import { NavbarItem, Button } from "@/components/nextui/client-components";
 import Link from "next/link";
 import { AccountMe } from "@/lib/api";
 import { cn } from "@/lib";
