+++ frontend/components/nextui/tabs.tsx
@@ -2,7 +2,7 @@
 
 import React from "react";
 
-import { Tabs as NextUiTabs, SlotsToClasses, Tab } from "@/components/nextui-use-client";
+import { Tabs as NextUiTabs, SlotsToClasses, Tab } from "@/components/nextui/client-components";
 import { useSearchParamsTabState } from "@/lib/hooks/use-search-params-tab-state";
 
 const tabs = ["dashboard", "profiles", "pokemon", "tournament_history", "settings", "admin"];
