+++ frontend/components/pokemon/pokemon-showdown-set-form.tsx
@@ -1,6 +1,6 @@
 import { postPokemonTeam } from "@/app/server-actions/pokemon/actions";
 import { ValidatedPokemon, PokePasteMetadata } from "@/lib/pokemon/common";
-import { Textarea, Spacer, Button } from "@/components/nextui-use-client";
+import { Textarea, Spacer, Button } from "@/components/nextui/client-components";
 import Form from "next/form";
 
 interface PokemonShowdownSetFormProps {
