+++ frontend/components/pokemon/pokemon-team-client-page.tsx
@@ -2,7 +2,7 @@
 
 import * as React from "react";
 import { usePokemonTeam } from "@/lib/pokemon/use-pokemon-team";
-import { Spacer } from "@/components/nextui-use-client";
+import { Spacer } from "@/components/nextui/client-components";
 import { PokemonShowdownSetForm } from "./pokemon-showdown-set-form";
 import { PokemonTeamDisplayGrid } from "./pokemon-team-display-grid";
