+++ frontend/components/profiles/new-profile.tsx
@@ -2,7 +2,7 @@
 import { useRouter } from "next/navigation";
 
 import React from "react";
-import { Button, Input } from "../nextui-use-client";
+import { Button, Input } from "../nextui/client-components";
 import { createProfile } from "@/app/server-actions/profiles/actions";
 import { AccountMe } from "@/lib/api";
 import Form from "next/form";
