+++ frontend/components/settings/theme-custom-radio.tsx
@@ -1,7 +1,7 @@
-import type { RadioProps } from "@/components/nextui-use-client";
+import type { RadioProps } from "@/components/nextui/client-components";
 
 import React from "react";
-import { useRadio, VisuallyHidden } from "@/components/nextui-use-client";
+import { useRadio, VisuallyHidden } from "@/components/nextui/client-components";
 
 import { cn } from "@/lib/utils";
