+++ frontend/lib/hooks/use-memoized-callback.ts
@@ -1,18 +1,17 @@
-import { useMemo, useRef } from "react";
+import * as React from "react";
 
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-type noop = (this: any, ...args: any[]) => any;
+type noop = (this: any, ...args: any[]) => any; // eslint-disable-line @typescript-eslint/no-explicit-any
 
 type PickFunction<T extends noop> = (this: ThisParameterType<T>, ...args: Parameters<T>) => ReturnType<T>;
 
 export function useMemoizedCallback<T extends noop>(fn: T) {
-  const fnRef = useRef<T>(fn);
+  const fnRef = React.useRef<T>(fn);
 
   // why not write `fnRef.current = fn`?
   // https://github.com/alibaba/hooks/issues/728
-  fnRef.current = useMemo<T>(() => fn, [fn]);
+  fnRef.current = React.useMemo<T>(() => fn, [fn]);
 
-  const memoizedFn = useRef<PickFunction<T>>(undefined);
+  const memoizedFn = React.useRef<PickFunction<T> | null>(null);
 
   if (!memoizedFn.current) {
     memoizedFn.current = function (this, ...args) {
